+++ TDEngine2/source/core/CProjectSettings.cpp
@@ -10,6 +10,7 @@ namespace TDEngine2
 		static const std::string mCommonSettingsGroupId;
 		static const std::string mGraphicsSettingsGroupId;
 		static const std::string mAudioSettingsGroupId;
+		static const std::string mLocalizationSettingsGroupId;
 
 		struct TCommonSettingsKeys
 		{
@@ -27,11 +28,21 @@ namespace TDEngine2
 		{
 			static const std::string mAudioTypeKey;
 		};
+
+		struct TLocalizationSettingsKeys
+		{
+			static const std::string mCurrActiveLocaleKey;
+			static const std::string mRegisteredLocalesKey;
+			static const std::string mLocaleInfoGroupId;
+			static const std::string mLocaleIdKey;
+			static const std::string mLocalePackagePathKey;
+		};
 	};
 
 	const std::string TProjectSettingsArchiveKeys::mCommonSettingsGroupId = "common_settings";
 	const std::string TProjectSettingsArchiveKeys::mGraphicsSettingsGroupId = "graphics_settings";
 	const std::string TProjectSettingsArchiveKeys::mAudioSettingsGroupId = "audio_settings";
+	const std::string TProjectSettingsArchiveKeys::mLocalizationSettingsGroupId = "localization_settings";
 
 	const std::string TProjectSettingsArchiveKeys::TCommonSettingsKeys::mApplicationIdKey = "application_id";
 	const std::string TProjectSettingsArchiveKeys::TCommonSettingsKeys::mMaxThreadsCountKey = "max_worker_threads_count";
@@ -40,6 +51,12 @@ namespace TDEngine2
 	const std::string TProjectSettingsArchiveKeys::TGraphicsSettingsKeys::mGraphicsTypeKey = "gapi_type";
 
 	const std::string TProjectSettingsArchiveKeys::TAudioSettingsKeys::mAudioTypeKey = "api_type";
+
+	const std::string TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mCurrActiveLocaleKey = "current_locale";
+	const std::string TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mRegisteredLocalesKey = "registered_locales";
+	const std::string TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mLocaleInfoGroupId = "locale_info";
+	const std::string TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mLocaleIdKey = "id";
+	const std::string TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mLocalePackagePathKey = "package_path";
 
 
 	CProjectSettings::CProjectSettings():
@@ -76,6 +93,38 @@ namespace TDEngine2
 		result = result | pFileReader->BeginGroup(TProjectSettingsArchiveKeys::mAudioSettingsGroupId);
 		{
 			 mAudioSettings.mAudioContextType = Meta::EnumTrait<E_AUDIO_CONTEXT_API_TYPE>::FromString(pFileReader->GetString(TProjectSettingsArchiveKeys::TAudioSettingsKeys::mAudioTypeKey));
+		}
+		result = result | pFileReader->EndGroup();
+
+		/// \note Localization settings
+		result = result | pFileReader->BeginGroup(TProjectSettingsArchiveKeys::mLocalizationSettingsGroupId);
+		{
+			pFileReader->BeginGroup(TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mRegisteredLocalesKey);
+			{
+				TLocalizationSettings::TLocaleInfo info;
+
+				auto& registeredLocales = mLocalizationSettings.mRegisteredLocales;
+
+				while (pFileReader->HasNextItem())
+				{
+					pFileReader->BeginGroup(Wrench::StringUtils::GetEmptyStr());
+					{
+						pFileReader->BeginGroup(TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mLocaleInfoGroupId);
+						
+						info.mName = pFileReader->GetString(TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mLocaleIdKey);
+						info.mPackagePath = pFileReader->GetString(TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mLocalePackagePathKey);
+						info.mId = static_cast<TLocaleId>(registeredLocales.size());
+
+						registeredLocales.emplace_back(info);
+
+						pFileReader->EndGroup();
+					}
+					pFileReader->EndGroup();
+				}
+			}
+			pFileReader->EndGroup();
+
+			mLocalizationSettings.mCurrActiveLocaleId = pFileReader->GetString(TProjectSettingsArchiveKeys::TLocalizationSettingsKeys::mCurrActiveLocaleKey);
 		}
 		result = result | pFileReader->EndGroup();